# Backend/src/Aurigma.DirectMail.Sample.TypeScriptApiClient/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Direct Mail API",
    "version": "v1"
  },
  "paths": {
...
    "/api/direct-mail/v1/line-items": {
      "get": {
        "tags": [
          "LineItem"
        ],
        "summary": "Returns a list of all line items.",
        "operationId": "LineItem_GetLineItems",
        "parameters": [
          {
            "name": "CampaignId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LineItemDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LineItemDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LineItemDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "LineItem"
        ],
        "summary": "Creates a new line item.",
        "operationId": "LineItem_CreateLineItem",
        "requestBody": {
          "description": "",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/LineItemCreationModel"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LineItemCreationModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LineItemCreationModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LineItemCreationModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LineItemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineItemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineItemDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/direct-mail/v1/line-items/{id}": {
      "get": {
        "tags": [
          "LineItem"
        ],
        "summary": "Returns a line item by ID.",
        "operationId": "LineItem_GetLineItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Line item id.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LineItemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineItemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineItemDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "LineItem"
        ],
        "summary": "Updates the line item.",
        "operationId": "LineItem_UpdateLineItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Line item ID.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Update model Aurigma.DirectMail.Sample.WebApi.Models.LineItem.LineItemUpdateModel",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/LineItemUpdateModel"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LineItemUpdateModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LineItemUpdateModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LineItemUpdateModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LineItemDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineItemDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineItemDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/direct-mail/v1/line-items/{id}/finish": {
      "post": {
        "tags": [
          "LineItem"
        ],
        "summary": "Process finish line item personalization.",
        "operationId": "LineItem_FinishPersonalizationLineItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Line item id.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConflictDto": {
        "type": "object",
        "properties": {
          "propertyName": {
            "type": "string",
            "nullable": true
          },
          "propertyValue": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LineItemCreationModel": {
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "templateId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LineItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "productId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "campaignId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "templateId": {
            "type": "string",
            "nullable": true
          },
          "productVariantId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "designId": {
            "type": "string",
            "description": "Personalized design id.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LineItemUpdateModel": {
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string",
            "description": "Campaign id.",
            "format": "uuid"
          },
          "quantity": {
            "type": "integer",
            "description": "Product quantity.",
            "format": "int32"
          },
          "productId": {
            "type": "string",
            "description": "Product id.",
            "format": "uuid"
          },
          "templateId": {
            "type": "string",
            "description": "Template id.",
            "nullable": true
          },
          "designId": {
            "type": "string",
            "description": "Personalized design id.",
            "nullable": true
          },
          "productVariantId": {
            "type": "integer",
            "description": "Selected product variant id.",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  },
...
}